requirements.txt
pytest
appium-python-client
playwright